requirements.txt
keyboard
elevenlabs
numpy
openai
pvrecorder
python-dotenv

